入力処理を別クラスで作成。mainクラスのループ条件の記述修正。
