+++ src/main/java/org/example/InputProcess.java
@@ -1,0 +1,89 @@
+package org.example;
+
+import java.util.InputMismatchException;
+import java.util.NoSuchElementException;
+import java.util.Scanner;
+
+public class InputProcess {
+
+  //変数宣言
+  private String operator;
+
+  //Scannerオブジェクトの作成
+  Scanner scanner = new Scanner(System.in);
+
+  //値の入力
+  public double inputValue() {
+    //変数宣言
+    double value = 0;
+
+    //入力処理
+    while (true) {
+      try {
+        System.out.println("数値の入力---------------");
+        value = scanner.nextDouble();
+        scanner.nextLine();//バッファのクリア
+        break;//ループを抜ける
+      } catch (InputMismatchException e) {
+        e.printStackTrace();
+        System.out.println("数値を入力してください---------------");
+        scanner.nextLine();//バッファのクリア
+      } catch (NoSuchElementException e) {
+        e.printStackTrace();
+        System.out.println("入力がありません---------------");
+        scanner.nextLine();//バッファのクリア
+      }
+    }
+    return value;
+  }
+
+  //演算子の入力
+  public String inputOperator() {
+    //入力処理
+    while (true) {
+      try {
+        System.out.println("演算子(+,-,*,/)を入力してください----------");
+        operator = scanner.nextLine();
+        if (!operator.equals("+") && !operator.equals("-") && !operator.equals("*")
+            && !operator.equals("/")) {
+          System.out.println("演算子が入力されていません----------");
+        } else {
+          break;
+        }
+      } catch (NoSuchElementException e) {
+        e.printStackTrace();
+        System.out.println("入力がありません---------------");
+      }
+    }
+    return operator;
+  }
+
+  //処理を継続するか確認
+  public boolean getUserContinue() {
+    //変数宣言
+    String userResponse;
+    boolean flag = true;
+    while (true) {
+      try {
+        System.out.println("処理を継続しますか？(yesまたはnoを入力)----------");
+        userResponse = scanner.nextLine();
+        switch (userResponse) {
+          case "yes":
+            break;
+          case "no":
+            System.out.println("=====終了=====");
+            flag = false;
+            break;
+          default:
+            System.out.println("無効な文字列です。再入力してください。----------");
+            continue;
+        }
+        break;
+      } catch (NoSuchElementException e) {
+        e.printStackTrace();
+        System.out.println("入力がありません---------------");
+      }
+    }
+    return flag;
+  }
+}
