+++ src/main/java/org/example/Main.java
@@ -12,11 +12,12 @@ public class Main {
 
     //オブジェクトのインスタンス化
    Calculator calculator = new Calculator();
+   InputProcess inputProcess = new InputProcess();
 
     //入力処理
-    value1 = calculator.inputValue();
-    operator = calculator.inputOperator();
-    value2 = calculator.inputValue();
+    value1 = inputProcess.inputValue();
+    operator = inputProcess.inputOperator();
+    value2 = inputProcess.inputValue();
 
     //フラグを立てる
     flag = true;
@@ -27,16 +28,16 @@ public class Main {
     System.out.println(result);
 
     //継続確認処理
-    flag = calculator.getUserContinue();
+    flag = inputProcess.getUserContinue();
 
-    while (flag == true) {
+    while (flag) {
 
       //計算結果の保存
       value1 = result;
 
       //入力処理
-      operator = calculator.inputOperator();
-      value2 = calculator.inputValue();
+      operator = inputProcess.inputOperator();
+      value2 = inputProcess.inputValue();
 
       //計算処理
       result = calculator.calculator(value1, operator, value2, flag);
@@ -44,7 +45,7 @@ public class Main {
       System.out.println(result);
 
       //継続確認処理
-      flag = calculator.getUserContinue();
+      flag = inputProcess.getUserContinue();
 
     }
   }
